+++ NEWS.md
@@ -7,6 +7,7 @@
 - geom_arc / stat_arc for drawing circle segments
 - geom_edge_bundle / stat_edge_bundle for drawing edge bundles based on control
 points
+- geom_arc_bar /stat_arc_bar / stat_pie for drawing arcs and wedges with fill
 - power_trans for creating power transformations
 - radial_trans for creating transformation between radial and cartesian 
 coordinates